requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Web scraping and data processing
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2

# NLP and ML
spacy==3.7.2
scikit-learn==1.3.2
nltk==3.8.1
pandas==2.1.3
numpy==1.25.2

# AI and text generation
openai==1.3.7
transformers==4.35.2
torch==2.1.1

# Database
sqlalchemy==1.4.53
aiosqlite==0.19.0

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1
